# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kartigo e-commerce platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/addresses:
    get:
      operationId: getUserAddresses
      tags: [users]
      summary: Get user addresses
      responses:
        "200":
          description: List of addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createAddress
      tags: [users]
      summary: Create a new address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Created address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /users/addresses/{id}:
    put:
      operationId: updateAddress
      tags: [users]
      summary: Update address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [users]
      summary: Delete address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: integer
        - name: rating
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}/helpful:
    post:
      operationId: markReviewHelpful
      tags: [reviews]
      summary: Mark review as helpful
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── SEARCH ────────────────────────────────────────────────────────────────
  /search:
    get:
      operationId: searchProducts
      tags: [search]
      summary: Search products
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: rating
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /search/suggestions:
    get:
      operationId: getSearchSuggestions
      tags: [search]
      summary: Get search autocomplete suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchSuggestions"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        replitId:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [BUYER, SELLER, ADMIN]
        rewardCoins:
          type: integer
        referralCode:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - replitId
        - role
        - rewardCoins
        - isActive
        - createdAt
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        avatar:
          type: string
    Address:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        landmark:
          type: ["string", "null"]
        isDefault:
          type: boolean
        type:
          type: string
          enum: [HOME, WORK, OTHER]
      required:
        - id
        - userId
        - fullName
        - phone
        - addressLine1
        - city
        - state
        - pincode
        - isDefault
        - type
    AddressInput:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        landmark:
          type: string
        isDefault:
          type: boolean
        type:
          type: string
          enum: [HOME, WORK, OTHER]
      required:
        - fullName
        - phone
        - addressLine1
        - city
        - state
        - pincode
        - type
...
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        basePrice:
          type: number
        sellingPrice:
          type: number
        discount:
          type: integer
        images:
          type: array
          items:
            type: string
        thumbnail:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        brandId:
          type: ["integer", "null"]
        brandName:
          type: ["string", "null"]
        sellerId:
          type: ["integer", "null"]
        sellerName:
          type: ["string", "null"]
        stock:
          type: integer
        averageRating:
          type: number
        totalReviews:
          type: integer
        totalSold:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        isFlashSale:
          type: boolean
        flashSalePrice:
          type: ["number", "null"]
        flashSaleEnd:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - slug
        - description
        - basePrice
        - sellingPrice
        - discount
        - images
        - thumbnail
        - categoryId
        - stock
        - averageRating
        - totalReviews
        - totalSold
        - isActive
        - isFeatured
        - isFlashSale
        - tags
        - createdAt
...
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages

    # ── CATEGORIES ───────────────────────────────────────────────────────────
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        productId:
          type: integer
        rating:
          type: integer
        title:
          type: ["string", "null"]
        body:
          type: string
        images:
          type: array
          items:
            type: string
        isVerifiedPurchase:
          type: boolean
        helpfulCount:
          type: integer
        isApproved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - userName
        - productId
        - rating
        - body
        - images
        - isVerifiedPurchase
        - helpfulCount
        - isApproved
        - createdAt
    ReviewInput:
      type: object
      properties:
        productId:
          type: integer
        orderId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
        images:
          type: array
          items:
            type: string
      required:
        - productId
        - rating
        - body
    ReviewBreakdown:
      type: object
      properties:
        average:
          type: number
        total:
          type: integer
        five:
          type: integer
        four:
          type: integer
        three:
          type: integer
        two:
          type: integer
        one:
          type: integer
      required:
        - average
        - total
        - five
        - four
        - three
        - two
        - one
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        breakdown:
          $ref: "#/components/schemas/ReviewBreakdown"
        total:
          type: integer
        page:
          type: integer
      required:
        - reviews
        - total
        - page

    # ── SEARCH ───────────────────────────────────────────────────────────────
    SearchSuggestions:
      type: object
      properties:
        products:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              slug:
                type: string
              thumbnail:
                type: string
              sellingPrice:
                type: number
            required:
              - id
              - title
              - slug
              - thumbnail
              - sellingPrice
        categories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              slug:
                type: string
            required:
              - name
              - slug
        brands:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              slug:
                type: string
            required:
              - name
              - slug
      required:
        - products
        - categories
        - brands

    # ── SELLERS ──────────────────────────────────────────────────────────────
...